requirements.txt
fastapi[standard]==0.115.6 
opencv-python==4.11.0.86
python-json-logger==3.2.1
python-dotenv==1.0.1
json2html==1.3.0
Jinja2==3.1.5
uvicorn==0.34.0